Abstand ins Programm integriert, Setup in Main getan
