+++ programs/Main/Main.ino
@@ -47,7 +47,7 @@
 
 void setup()
 {
-  delay(300);
+  delay(5000);                       // Wichtig für den Abstandssensor
   pinMode(LED_BUILTIN, OUTPUT);      // Pin D13
   pinMode(motorpin, INPUT_PULLDOWN); // define pinmode for switch on the side of the bot
   pinMode(kalibrierung, INPUT);      // define pinmode for calibration button
@@ -86,6 +86,19 @@ void setup()
   sensorLeiste.setTypeRC();
   sensorLeiste.setSensorPins(SENSOR_LEISTE_PINS, SENSOR_LEISTE_ANZAHL_SENSOREN);
   Serial.println("Initialisierung Reflektionssensor abgeschlossen");
+
+  // ABSTANDSSENSOR-INITIALISIEREN
+  if (!abstandsSensor.begin(NEUE_ABSTANDSADDRESSE, Wire)) { // Fisch: Die dinger werden ausgeführt und dabei nach Fehlern getestet
+    delay(10000); // damit wir Zeit haben den Serial Monitor zu öffnen nach dem Upload
+    Serial.println("ToF64 Verdrahtung prüfen! Roboter aus- und einschalten! Programm Ende.");
+    while (1);
+  }
+  if (!abstandsSensor.setResolution(einstellungen.aufloesung) ||
+    !abstandsSensor.setRangingFrequency(einstellungen.maxMessfrequenz)) {  // siehe oben
+      delay(10000); // damit wir Zeit haben den Serial Monitor zu öffnen nach dem Upload
+      Serial.println("ToF64 Auflösung oder Messfrequenz konnte nicht geändert werden! Programm Ende.");
+      while (1);
+  }
   motors.initialize();
   // falls man global die Motor-Drehrichtung ändern möchte:
   motors.flipLeftMotor(false); // nur notwendig, wenn man true reinschreibt
@@ -97,6 +110,8 @@ void setup()
 #include "Farbauslese.h"        //commands for reading and processing colorsensors
 #include "doppelschwarz.h"      //command for handling crosssections
 #include "Opfer.h"              //Du Opfer
+
+#include "Abstand.h"            // Abstand, noch nicht einsortiert zwischen die restlichen includes
 
 int x = 0;
 int y = 0;
