+++ programs/Main/variables.h
@@ -31,21 +31,31 @@ RescueBoardMotors motors = RescueBoardMotors();
 //ABSTANDSSENSOR
 SparkFun_VL53L5CX abstandsSensor = SparkFun_VL53L5CX();
 const uint8_t NEUE_ABSTANDSADDRESSE = 0x35;
-typedef struct Einstellungen {
+bool debT = false; // debT = debug timer, machen wir an falls wir Probleme haben und dann sagt der Timer ob überhaupt Werte ankommen
+typedef struct Einstellungen { // Structs sind schön weil die so wie Listen mit Konstruktor und verschiedenen Datentypen sind
     uint8_t aufloesung;
     uint8_t bildSeitenlaenge;
     uint8_t maxMessfrequenz;
-} Einstellungen_t;
-// Bei meinem Sensor habe ich hier ein paar "kaputte Pixel":
+} Einstellungen_t; // null ahnung was das eigentlich macht
+// Könnte "kaputte Pixel" ausgeben:
 const Einstellungen ACHT_MAL_ACHT = { VL53L5CX_RESOLUTION_8X8, 8, 15 }; // 8x8: max 15Hz
-// Bei meinem Sensor sieht hier alles gut aus:
+// Vermutlich stabiler
 const Einstellungen VIER_MAL_VIER = { VL53L5CX_RESOLUTION_4X4, 4, 60 }; // 4x4: max 60Hz
 Einstellungen einstellungen = VIER_MAL_VIER;
+enum Modus {
+    /* Werte im Serial Monitor anzeigen. */
+    ABSTANDS_WERTE_LOGGEN,
+    /* Sensor gehen eine Fläche richten (z.B. Monitor) und in weniger als SCHWELLENWERT mm etwas davor halten.  */
+    SCHWELLENWERT_VISUALISIERUNG,
+};
+const int SCHWELLENWERT = 100;
+/** hier einstellen, was das Programm mit den Sensorwerten anfangen soll: */
+Modus modus = SCHWELLENWERT_VISUALISIERUNG;
 // hier speichern wir die 6 TOFsensorwerte ab:
 VL53L5CX_ResultsData messDaten;
 
 
-//IRGENDEINE SCHEIẞE
+//IRGENDWAS
 #define CMPS12 0x60
 uint16_t direction;
 uint16_t currentdirection;
